Title: a의 b승, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240818/a의 b승/a-to-the-power-of-b.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int a,b, result =1;
+    long long a,b, result =1;
     cin >> a >> b;
     result = a;
     for(int i = 1; i <b; i++)
